Authentification et création de compte

+++ www/js/controlers/connectionMailCtrls.js
@@ -16,12 +16,20 @@ starter
         OnAuthenticateError(data);
         return;
       }
+      data = data[0]['value'];
       console.log(data);
+
+      data = JSON.parse(data);
+      if(data.length ==0){
+        OnAuthenticateError(data);
+        return; 
+      }
+
       localStorageService.remove('connexion');
       var connexion = {
         'etat': true,
         'libelle': 'Se déconnecter',
-        'employeID': data.employerId
+        'employeID': data.jobyerId
       };
 
       localStorageService.set('connexion', connexion);
